+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Removed use of the deprecated Apex parser `CaseInsensitiveInputStream` from JVM parsing (#489)
+
 ## [6.1.0-beta.1] - 2026-06-13
 
 ### Added